docs: Update json editor tutorial 📚
