+++ src/tutorial/json-editor/README.md
@@ -1,26 +1,28 @@
 # JSON Editor
 
-In this tutorial, we will be creating an application that gives the user a simple interface to enter
-key-value pairs, which will be converted and printed to `stdout` in json.
+Now that we have covered some of the basics of a "hello world" and "counter" app, we are ready to
+build and manage something more involved.
 
-The purpose of this application will be to give the user an interface to create correct json,
-instead of having to worry about commas and brackets themselves.
+In this tutorial, we will be creating an application that gives the user a simple interface to enter
+key-value pairs, which will be converted and printed to `stdout` in json. The purpose of this
+application will be to give the user an interface to create correct json, instead of having to worry
+about commas and brackets themselves.
 
 ## Initialization
 
 Go ahead and set up a new rust project with
 
 ```sh
-cargo init ratatui-example-project
+cargo init ratatui-json-editor
 ```
 
 and put the following in the `Cargo.toml`:
 
 ```
-{{#include ../../../ratatui-book-tutorial-project/Cargo.toml:8:}}
+{{#include ./ratatui-json-editor-app/Cargo.toml:8:}}
 ```
 
-or the latest version of these libraries
+or the latest version of these libraries.
 
 ## Filestructure
