+++ src/tutorial/json-editor/app.md
@@ -1,8 +1,8 @@
 # App.rs
 
-As we saw in the previous section, a common model for smaller `ratatui` applications, is to have one
-large application state struct called `App` or some variant of that name. We will be using this
-paradigm in this application.
+As we saw in the previous section, a common model for smaller `ratatui` applications is to have one
+application state struct called `App` or some variant of that name. We will be using this paradigm
+in this application as well.
 
 This struct will contain all of our "persistent" data and will be passed to any function that needs
 to know the current state of the application.
@@ -11,12 +11,17 @@ to know the current state of the application.
 
 ## Application modes
 
-Often it is beneficial to imagine that there are several 'modes' the application can be in. Thinking
-this way will make it easier to segregate everything from what window is getting drawn, to what
-keybinds to listen for. We will be using the application's state and enums to track two things. The
-first thing we are going to track is what screen the user should be seeing, and the second will be
-an optional enum that only applies while the user is editing a key-value pair that tracks which (key
-or value) should be highlighted for the user.
+It is useful to think about the several "modes" that your application can be in. Thinking in "modes"
+will make it easier to segregate everything from what window is getting drawn, to what keybinds to
+listen for.
+
+We will be using the application's state to track two things:
+
+1. what screen the user is seeing,
+2. which box should be highlighted, the "key" or "value" (this only applies when the user is editing
+   a key-value pair).
+
+### Current Screen Enum
 
 In this tutorial application, we will have three "screens":
 
@@ -28,22 +33,23 @@ In this tutorial application, we will have three "screens":
 We represent these possible modes with a simple enum:
 
 ```rust,no_run,noplayground
-{{#include ../../../ratatui-book-tutorial-project/src/app.rs:screen_modes}}
+{{#include ./ratatui-json-editor-app/src/app.rs:screen_modes}}
 ```
 
-## Other state enums
+### Currently Editing Enum
 
-`Ratatui` does not automatically redraw the screen (See [Rendering](./../concepts/rendering.md) for
-more information), and it does not remember anything about what it drew last frame. This means that
-the programmer is responsible for handling all state and updating widgets to reflect changes. In
-this case, we will allow the user to input two strings in the `Editing` mode - a key and a value.
-The programmer is responsible for knowing which the user is trying to edit.
+As you may already know, `ratatui` does not automatically redraw the screen[^note]. `ratatui` also
+does not remember anything about what it drew last frame.
+
+This means that the programmer is responsible for handling all state and updating widgets to reflect
+changes. In this case, we will allow the user to input two strings in the `Editing` mode - a key and
+a value. The programmer is responsible for knowing which the user is trying to edit.
 
 For this purpose, we will create another enum for our application state called `CurrentlyEditing` to
 keep track of which field the user is currently entering:
 
 ```rust,no_run,noplayground
-{{#include ../../../ratatui-book-tutorial-project/src/app.rs:currently_editing}}
+{{#include ./ratatui-json-editor-app/src/app.rs:currently_editing}}
 ```
 
 ## The full application state
@@ -52,7 +58,7 @@ Now that we have enums to help us track where the user is, we will create the st
 stores this data which can be passed around where it is needed.
 
 ```rust,no_run,noplayground
-{{#include ../../../ratatui-book-tutorial-project/src/app.rs:app_fields}}
+{{#include ./ratatui-json-editor-app/src/app.rs:app_fields}}
 ```
 
 ## Helper functions
@@ -61,18 +67,18 @@ While we could simply keep our application state as simply a holder of values, w
 few helper functions which will make our life easier elsewhere. Of course, these functions should
 only affect the application state itself, and nothing outside of it.
 
-#### new()
+#### `new()`
 
 We will be adding this function simply to make creating the state easier. While this could be
 avoided by specifying it all in the instantiation of the variable, doing it here allows for easy to
 change universal defaults for the state.
 
 ```rust,no_run,noplayground
-{{#include ../../../ratatui-book-tutorial-project/src/app.rs:impl_new}}
+{{#include ./ratatui-json-editor-app/src/app.rs:impl_new}}
     ...
 ```
 
-#### save_key_value()
+#### `save_key_value()`
 
 This function will be called when the user saves a key-value pair in the editor. It adds the two
 stored variables to the key-value pairs `HashMap`, and resets the status of all of the editing
@@ -80,11 +86,11 @@ variables.
 
 ```rust,no_run,noplayground
     ...
-{{#include ../../../ratatui-book-tutorial-project/src/app.rs:save_key_value}}
+{{#include ./ratatui-json-editor-app/src/app.rs:save_key_value}}
     ...
 ```
 
-#### toggle_editing()
+#### `toggle_editing()`
 
 Sometimes it is easier to put simple logic into a convenience function so we don't have to worry
 about it in the main code block. `toggle_editing` is one of those cases. All we are doing, is
@@ -93,17 +99,20 @@ Value fields.
 
 ```rust,no_run,noplayground
     ...
-{{#include ../../../ratatui-book-tutorial-project/src/app.rs:toggle_editing}}
+{{#include ./ratatui-json-editor-app/src/app.rs:toggle_editing}}
     ...
 ```
 
-#### print_json()
+#### `print_json()`
 
 Finally, is another convenience function to print out the serialized json from all of our key-value
 pairs.
 
 ```rust,no_run,noplayground
     ...
-{{#include ../../../ratatui-book-tutorial-project/src/app.rs:print_json}}
+{{#include ./ratatui-json-editor-app/src/app.rs:print_json}}
     ...
 ```
+
+<!-- prettier-ignore -->
+[^note]: In ratatui, every frame draws the UI anew. See the [Rendering section](./../../concepts/rendering.md) for more information.
