+++ src/tutorial/json-editor/ratatui-json-editor-app/.rustfmt.toml
@@ -1,0 +1,1 @@
+tab_spaces = 4
