+++ src/tutorial/json-editor/ratatui-json-editor-app/shell.nix
