+++ src/tutorial/json-editor/ratatui-json-editor-app/src/app.rs
@@ -22,8 +22,8 @@ pub enum CurrentlyEditing {
 
 // ANCHOR: app_fields
 pub struct App {
-    pub key_input: String, // the currently being edited json key.
-    pub value_input: String, // the currently being edited json value.
+    pub key_input: String,              // the currently being edited json key.
+    pub value_input: String,            // the currently being edited json value.
     pub pairs: HashMap<String, String>, // The representation of our key and value pairs with serde Serialize support
     pub current_screen: CurrentScreen, // the current screen the user is looking at, and will later determine what is rendered.
     pub currently_editing: Option<CurrentlyEditing>, // the optional state containing which of the key or value pair the user is editing. It is an option, because when the user is not directly editing a key-value pair, this will be set to `None`.
@@ -41,12 +41,12 @@ impl App {
             currently_editing: None,
         }
     }
-// ANCHOR_END: impl_new
+    // ANCHOR_END: impl_new
 
     // ANCHOR: save_key_value
     pub fn save_key_value(&mut self) {
-
-        self.pairs.insert(self.key_input.clone(), self.value_input.clone());
+        self.pairs
+            .insert(self.key_input.clone(), self.value_input.clone());
 
         self.key_input = String::new();
         self.value_input = String::new();
