+++ src/tutorial/json-editor/ratatui-json-editor-app/src/ui.rs
@@ -4,7 +4,7 @@ use ratatui::{
     layout::{Constraint, Direction, Layout, Rect},
     style::{Color, Style},
     text::{Line, Span, Text},
-    widgets::{Block, Borders, Clear, Paragraph, Wrap, ListItem, List},
+    widgets::{Block, Borders, Clear, List, ListItem, Paragraph, Wrap},
     Frame,
 };
 
@@ -12,7 +12,7 @@ use crate::app::{App, CurrentScreen, CurrentlyEditing};
 
 // ANCHOR: method_sig
 pub fn ui<B: Backend>(f: &mut Frame<B>, app: &App) {
-// ANCHOR_END: method_sig
+    // ANCHOR_END: method_sig
     // Create the layout sections.
     // ANCHOR: ui_layout
     let chunks = Layout::default()
@@ -45,16 +45,10 @@ pub fn ui<B: Backend>(f: &mut Frame<B>, app: &App) {
     let mut list_items = Vec::<ListItem>::new();
 
     for key in app.pairs.keys() {
-        list_items.push(
-            ListItem::new(
-                Line::from(
-                    Span::styled(
-                        format!("{: <25} : {}", key, app.pairs.get(key).unwrap()),
-                        Style::default().fg(Color::Yellow),
-                    )
-                )
-            )
-        );
+        list_items.push(ListItem::new(Line::from(Span::styled(
+            format!("{: <25} : {}", key, app.pairs.get(key).unwrap()),
+            Style::default().fg(Color::Yellow),
+        ))));
     }
 
     let list = List::new(list_items);
@@ -65,18 +59,25 @@ pub fn ui<B: Backend>(f: &mut Frame<B>, app: &App) {
     let current_navigation_text = vec![
         // The first half of the text
         match app.current_screen {
-            CurrentScreen::Main    => Span::styled("Normal Mode", Style::default().fg(Color::Green)),
-            CurrentScreen::Editing => Span::styled("Editing Mode", Style::default().fg(Color::Yellow)),
+            CurrentScreen::Main => Span::styled("Normal Mode", Style::default().fg(Color::Green)),
+            CurrentScreen::Editing => {
+                Span::styled("Editing Mode", Style::default().fg(Color::Yellow))
+            }
             CurrentScreen::Exiting => Span::styled("Exiting", Style::default().fg(Color::LightRed)),
-        }.to_owned(),
+        }
+        .to_owned(),
         // A white divider bar to separate the two sections
         Span::styled(" | ", Style::default().fg(Color::White)),
         // The final section of the text, with hints on what the user is editing
         {
             if let Some(editing) = &app.currently_editing {
                 match editing {
-                    CurrentlyEditing::Key => Span::styled("Editing Json Key", Style::default().fg(Color::Green)),
-                    CurrentlyEditing::Value => Span::styled("Editing Json Value", Style::default().fg(Color::LightGreen)),
+                    CurrentlyEditing::Key => {
+                        Span::styled("Editing Json Key", Style::default().fg(Color::Green))
+                    }
+                    CurrentlyEditing::Value => {
+                        Span::styled("Editing Json Value", Style::default().fg(Color::LightGreen))
+                    }
                 }
             } else {
                 Span::styled("Not Editing Anything", Style::default().fg(Color::DarkGray))
@@ -210,8 +211,7 @@ fn centered_rect(percent_x: u16, percent_y: u16, r: Rect) -> Rect {
             ]
             .as_ref(),
         )
-        .split(popup_layout[1])
-        [1] // Return the middle chunk
+        .split(popup_layout[1])[1] // Return the middle chunk
 }
 // ANCHOR_END: centered_rect
